note/8_分笔数据实时订阅与日线采样/增强版v2.0/windows_端/enhanced_requirements.txt
# QMT增强版数据订阅系统依赖包
# 基于最佳实践的高性能依赖配置

# 核心依赖
xtquant>=1.7.0              # QMT量化交易接口
redis>=4.5.0                # Redis客户端
PyYAML>=6.0                 # YAML配置文件解析

# 数据处理
pandas>=1.5.0               # 数据分析库
numpy>=1.24.0               # 数值计算库

# 系统监控
psutil>=5.9.0               # 系统资源监控

# 网络和并发
requests>=2.28.0            # HTTP请求库
aioredis>=2.0.0             # 异步Redis客户端 (可选)

# 日志和调试
colorlog>=6.7.0             # 彩色日志输出 (可选)

# 性能优化 (可选)
ujson>=5.7.0                # 高性能JSON处理
msgpack>=1.0.0              # 高效序列化格式
lz4>=4.3.0                  # 快速压缩算法

# 开发和测试 (可选)
pytest>=7.2.0              # 测试框架
pytest-cov>=4.0.0          # 测试覆盖率
black>=23.0.0               # 代码格式化
flake8>=6.0.0               # 代码检查
